Add files via upload
Titanic survivor prediction is one of most famous problems for the beginners who are learning machine learning. In this repository, one can find data cleaning and data analysis of dataset. I tried and tested various algorithms to check which one gave me the best prediction accuracy and used it in the final code.